Change defaults to work out of the box with vagrant
Holy shit I actually got it to work.

+++ install.sh
@@ -2,13 +2,16 @@
 # This should be ran from inside the r/a/dio vagrant box:
 # https://r-a-d.io/vagrant.box
 
-composer self-update
+if [ ! -f app/config/app.php ]; then
+	cp app/config/app.php.sample app/config/app.php
+fi
 
-cp app/config/app.php.sample app/config/app.php
-cp app/config/database.php.sample app/config/database.php
+if [ ! -f app/config.database.php ]; then
+	cp app/config/database.php.sample app/config/database.php
+fi
 
 echo "Edit app/config/app.php and app/config/database.php after this."
 
 composer install
 
-# Done.
+# done
